Selvach/refactor constants (#505)
* Refactor Constants to all be in appConfig

* Move appConfig within src and switch to module imports
