--- src/features/register/constants.ts
@@ -1,9 +1,0 @@
-export const NursesConsentVersionUS = 'v2';
-export const consentVersionUS = 'v1';
-export const consentVersionUK = 'v2';
-export const consentVersionSE = 'v1';
-export const privacyPolicyVersionUS = 'v1.3';
-export const privacyPolicyVersionUK = 'v2.3';
-export const privacyPolicyVersionSE = 'v2.1';
-export const ukValidationStudyConsentVersion = 'v3';
-export const ukValidationStudyAdVersion = 'v2';
